Move lens utils to their own module
